bin/modules/commerce-services/commerceservices/testsrc/de/hybris/platform/commerceservices/order/validator/impl/FallbackAddToCartValidatorTest.java
/*
 * Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved.
 */
package de.hybris.platform.commerceservices.order.validator.impl;

import de.hybris.bootstrap.annotations.UnitTest;
import de.hybris.platform.commerceservices.order.CommerceCartModificationException;
import de.hybris.platform.commerceservices.service.data.CommerceCartParameter;
import de.hybris.platform.core.model.product.ProductModel;
import de.hybris.platform.variants.model.VariantTypeModel;

import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;

import static org.junit.Assert.assertTrue;


/**
 * Units tests for {@link FallbackAddToCartValidator}
 */
@UnitTest
public class FallbackAddToCartValidatorTest
{
	private final FallbackAddToCartValidator classUnderTest = new FallbackAddToCartValidator();
	private final CommerceCartParameter parameter = new CommerceCartParameter();
	@Mock
	private ProductModel product;
	@Mock
	private VariantTypeModel variantType;

	@Before
	public void initialize()
	{
		MockitoAnnotations.initMocks(this);
	}

	@Test
	public void testSupports()
	{
		assertTrue(classUnderTest.supports(parameter));
	}

	@Test(expected = IllegalArgumentException.class)
	public void testValidateNull() throws CommerceCartModificationException
	{
		classUnderTest.validate(null);
	}

	@Test(expected = IllegalArgumentException.class)
	public void testValidateNoProduct() throws CommerceCartModificationException
	{
		classUnderTest.validate(parameter);
	}

	@Test
	public void testValidateProductValid() throws CommerceCartModificationException
	{
		parameter.setProduct(product);
		classUnderTest.validate(parameter);
	}

	@Test(expected = CommerceCartModificationException.class)
	public void testValidateProductVariantBaseProduct() throws CommerceCartModificationException
	{
		parameter.setProduct(product);
		Mockito.when(product.getVariantType()).thenReturn(variantType);
		classUnderTest.validate(parameter);
	}
}
